# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation - Archiv",
    "version": "1.0.0",
    "description": "Documentation de l'API pour le projet Archiv (Authentification et Projets)"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement local"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth_token"
      }
    }
  },
  "paths": {
...
    "/api/projects/{id}": {
      "put": {
        "summary": "Modifier un projet",
        "tags": [
          "Projets"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "pdf": {
                    "type": "string",
                    "format": "binary"
                  },
                  "cover": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Projet mis à jour"
          },
          "403": {
            "description": "Non autorisé à modifier ce projet"
          }
        }
      },
      "delete": {
        "summary": "Supprimer un projet",
        "tags": [
          "Projets"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projet supprimé"
          },
          "403": {
            "description": "Non autorisé à supprimer ce projet"
          }
        }
      }
    },
...
    "/api/admin/projects/{id}": {
      "delete": {
        "summary": "Supprimer un projet (Admin)",
        "tags": ["Admin"],
        "security": [{"cookieAuth": []}],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Projet supprimé" }
        }
      }
    }
  }
}